requirements.txt
opencv-python
numpy
ultralytics
pyserial 